Dashboard and Display user blogs merged
