+++ frontend/src/Final/DisplayBlogs.css
@@ -102,7 +102,6 @@
 .about-user,
 .avatar-blogs {
     display: inline-block;
-    position: relative;
 }
 
 .about-user {
